Improve token example
