+++ x/programs/rust/examples/token/Cargo.toml
@@ -15,6 +15,10 @@ wasmlanche_sdk = { version = "0.1.0", path = "../../wasmlanche_sdk", features = 
 serde_json = "1.0.68"
 serial_test = "2.0.0"
 
+[[test]]
+name = "integration"
+path = "tests/examples.rs"
+
 [lib]
 crate-type = ["cdylib"] # set the crate(needed for cargo build to work properly)
